Introduce gammapy.cube.models; move MapEvaluator with MapFit
