+++ gammapy/cube/__init__.py
@@ -2,7 +2,6 @@
 """
 Sky cubes (3-dimensional: energy, lon, lat).
 """
-from .models import *
 from .counts import *
 from .exposure import *
 from .background import *
